pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta:__legacy__"

[project]
name = "world-codespaces-api-client"
version = "0.1.0"
dependencies = [
    "arglite",
    "pennant",
    "pyyaml",
    "fake-useragent",
    "python-dotenv",
    "requests",
    "rich"
]

readme = "README.md"
description = "Client-sided API interaction installed to Codespaces container."

[project.scripts]
who = "omnipresence:Who.cmd"
talk = "persona:Talk.cmd"
look = "persona:Look.cmd"
presence = "omnipresence:main.report"
inventory = "inventory:main.list"
get = "inventory:Acquire.cmd"
use = "inventory:Use.cmd_use"
give = "inventory:Give.cmd"
info = "inventory:Use.cmd_info"
drop = "inventory:Drop.cmd"
climate = "climate:main.main"
